revert changes that shouldn't end up in base layer
remove elasticsearch client interface

added blocked status when java not related

add positional arg to install_elasticserch func
