+++ layer.yaml
@@ -2,5 +2,4 @@ includes:
   - 'layer:basic'
   - 'layer:apt'
   - 'interface:java'
-  - 'interface:elasticsearch'
 repo: "https://github.com/jamesbeedy/juju-layer-elasticsearch-base"
