+++ metadata.yaml
@@ -13,7 +13,5 @@ series:
 - xenial
 subordinate: false
 provides:
-  client:
-    interface: elasticsearch
   java:
     interface: java
