+++ reactive/elasticsearch_base.py
@@ -1,35 +1,36 @@
 #!/usr/bin/env python3
 # pylint: disable=c0111,c0103,c0301
+import os
 import subprocess as sp
 
-from subprocess import CalledProcessError
-from jujubigdata import utils
 from charms.reactive import (
     when,
     when_not,
-    when_file_changed,
-    set_state,
-    remove_state
+    set_state
 )
 from charmhelpers.core.hookenv import (
     status_set,
-    application_version_set,
-    config,
-    open_port
+    application_version_set
 )
 from charmhelpers.core.host import (
     service_running,
-    service_start,
-    service_restart
+    service_start
 )
-from charms.layer.elasticsearch_base import (#pylint: disable=E0611,E0401,C0412
+from charms.layer.elasticsearch_base import (
+    # pylint: disable=E0611,E0401,C0412
     is_container,
     es_version
 )
 
-#@when('java.ready')
+
+@when_not('java.ready')
+def java_waiting_status():
+    status_set('blocked', 'Need relation to java providing application')
+
+
+@when('java.ready')
 @when_not('elasticsearch.installed')
-def install_elasticsearch():
+def install_elasticsearch(java):
     """Check for container, install elasticsearch
     """
     # Workaround for container installs is to set
@@ -37,37 +38,14 @@ def install_elasticsearch():
     # so kernel files will not need to be modified on
     # elasticsearch install. See
     # https://github.com/elastic/elasticsearch/commit/32df032c5944326e351a7910a877d1992563f791
-    # setting a environment variable will need to be exported
-    # otherwise it's only accesible in a new session
     if is_container():
-        with utils.environment_edit_in_place('/etc/environment') as env:
-            env['ES_SKIP_SET_KERNEL_PARAMETERS'] = 'true'
-        sp.check_call(['export', 'ES_SKIP_SET_KERNEL_PARAMETERS=true'], shell=True)
+        os.environ['ES_SKIP_SET_KERNEL_PARAMETERS'] = 'true'
     sp.call(['apt', 'install', 'elasticsearch', '-y',
              '--allow-unauthenticated'], shell=False)
     set_state('elasticsearch.installed')
 
+
 @when('elasticsearch.installed')
-@when_not('elasticsearch.configured')
-def configure_elasticsearch():
-    status_set('maintenance', 'Configuring elasticsearch')
-    # check if Firewall has to be enabled
-    init_fw()
-    set_state('elasticsearch.configured')
-    restart()
-
-@when('elasticsearch.configured')
-@when_file_changed('/etc/elasticsearch/elasticsearch.yml')
-def restart():
-    try:
-        status_set('maintenance', 'Restarting elasticsearch')
-        service_restart('elasticsearch')
-        set_state('elasticsearch.ready')
-        status_set('active', 'Ready')
-    except CalledProcessError:
-        status_set('error', 'Could not restart elasticsearch')
-
-@when('elasticsearch.ready')
 @when_not('elasticsearch.running')
 def ensure_elasticsearch_running():
     """Ensure elasticsearch is started
@@ -85,7 +63,7 @@ def ensure_elasticsearch_running():
         set_state('elasticsearch.problems')
 
 
-@when('elasticsearch.ready', 'elasticsearch.running')
+@when('elasticsearch.installed', 'elasticsearch.running')
 @when_not('elasticsearch.version.set')
 def get_set_elasticsearch_version():
     """Wait until we have the version to confirm
@@ -94,7 +72,6 @@ def get_set_elasticsearch_version():
 
     status_set('maintenance', 'Waiting for Elasticsearch to start')
     application_version_set(es_version())
-    open_port(9200)
     status_set('active', 'Elasticsearch started')
     set_state('elasticsearch.version.set')
 
@@ -115,52 +92,3 @@ def blocked_due_to_problems():
     """
     status_set('blocked',
                "There are problems with elasticsearch, please debug")
-
-######################
-# Relation functions #
-######################
-
-@when('client.connected', 'elasticsearch.configured')
-def connect_to_client(connected_clients):
-    conf = config()
-    cluster_name = conf['cluster-name']
-    port = conf['port']
-    connected_clients.configure(port, cluster_name)
-    clients = connected_clients.list_connected_clients_data
-    for c in clients:
-        add_fw_exception(c)
-
-@when('client.broken')
-def remove_client(broken_clients):
-    #
-    clients = broken_clients.list_connected_clients_data
-    for c in clients:
-        if c is not None:
-            rm_fw_exception(c)
-    remove_state('client.broken')
-
-######################
-# Firewall Functions #
-######################
-
-def init_fw():
-    conf = config()
-    #this value has te be changed to set ufw rules
-    utils.re_edit_in_place('/etc/default/ufw', {
-        r'IPV6=yes': 'IPV6=no',
-    })
-    if conf['firewall_enabled']:
-        sp.check_call(['ufw', 'allow', '22'])
-        sp.check_output(['ufw', 'enable'], input='y\n', universal_newlines=True)
-    else:
-        sp.check_output(['ufw', 'disable'])
-
-def add_fw_exception(host_ip):
-    sp.check_call([
-        'ufw', 'allow', 'proto', 'tcp', 'from', host_ip,
-        'to', 'any', 'port', '9200'])
-
-def  rm_fw_exception(host_ip):
-    sp.check_call([
-        'ufw', 'delete', 'allow', 'proto', 'tcp', 'from', host_ip,
-        'to', 'any', 'port', '9200'])
